Split Process/ProcessUtil. Process is platform dependant. ProcessUtil are functions that are platform independent.
