+++ source/core/slang-process-util.h
@@ -2,53 +2,14 @@
 #ifndef SLANG_PROCESS_UTIL_H
 #define SLANG_PROCESS_UTIL_H
 
-#include "slang-string.h"
-#include "slang-list.h"
-#include "slang-stream.h"
-
-#include "slang-string-escape-util.h"
+#include "slang-process.h"
 
 namespace Slang {
 
-struct CommandLine
-{
-    enum class ExecutableType
-    {
-        Unknown,                    ///< The executable is not specified 
-        Path,                       ///< The executable is set as a path
-        Filename,                   ///< The executable is set as a filename
-    };
-
-        /// Add args - assumed unescaped
-    void addArg(const String& in) { m_args.add(in); }
-    void addArgs(const String* args, Int argsCount) { for (Int i = 0; i < argsCount; ++i) addArg(args[i]); }
-
-        /// Find the index of an arg which is exact match for slice
-    SLANG_INLINE Index findArgIndex(const UnownedStringSlice& slice) const { return m_args.indexOf(slice); }
-
-        /// Set the executable path
-    void setExecutablePath(const String& path) { m_executableType = ExecutableType::Path; m_executable = path; }
-    void setExecutableFilename(const String& filename) { m_executableType = ExecutableType::Filename; m_executable = filename; }
-
-        /// For handling args where the switch is placed directly in front of the path 
-    SLANG_INLINE void addPrefixPathArg(const char* prefix, const String& path, const char* pathPostfix = nullptr);
-
-        /// Get the total number of args
-    SLANG_FORCE_INLINE Index getArgCount() const { return m_args.getCount(); }
-
-        /// Reset to the initial state
-    void reset() { *this = CommandLine();  }
-
-        /// Ctor
-    CommandLine():m_executableType(ExecutableType::Unknown) {}
-
-    ExecutableType m_executableType;    ///< How the executable is specified
-    String m_executable;                ///< Executable to run. Note that the executable is never escaped.
-    List<String> m_args;                ///< The arguments (Stored *unescaped*)
-};
-
 struct ExecuteResult
 {
+    typedef int ResultCode;
+
     void init()
     {
         resultCode = 0;
@@ -56,55 +17,13 @@ struct ExecuteResult
         standardError = String();
     }
 
-    typedef int ResultCode;
     ResultCode resultCode;
-    Slang::String standardOutput;
-    Slang::String standardError;
-};
-
-class Process : public RefObject
-{
-public:
-    enum class StreamType
-    {
-        ErrorOut,
-        StdOut,
-        StdIn,
-        CountOf,
-    };
-
-    typedef uint32_t Flags;
-    struct Flag
-    {
-        enum Enum : Flags
-        {
-            AttachDebugger = 0x01,
-        };
-    };
-
-        /// Get the stream for the type
-    Stream* getStream(StreamType type) const { return m_streams[Index(type)]; }
-    int32_t getReturnValue() const { return m_returnValue;  }
-
-        /// True if the process has terminated
-    virtual bool isTerminated() = 0;
-        /// Blocks until the process has completed
-    virtual void waitForTermination() = 0;
-
-protected:
-    int32_t m_returnValue = 0;                              ///< Value returned if process terminated
-    RefPtr<Stream> m_streams[Index(StreamType::CountOf)];   ///< Streams to communicate with the process
+    String standardOutput;
+    String standardError;
 };
 
 struct ProcessUtil
 {
-        /// The quoting style used for the command line on this target. Currently just uses Space,
-        /// but in future may take into account platform sec
-    static StringEscapeHandler* getEscapeHandler();
-
-        /// Get the suffix used on this platform
-    static UnownedStringSlice getExecutableSuffix();
-
         /// Output how the command line is executed on the target (with escaping and the such like)
     static String getCommandLineString(const CommandLine& commandLine);
 
@@ -117,32 +36,7 @@ struct ProcessUtil
 
         /// Read streams from process. 
     static SlangResult readUntilTermination(Process* process, ExecuteResult& outExecuteResult);
-
-        /// Create a process using the executable/args defined from the commandLine
-    static SlangResult createProcess(const CommandLine& commandLine, Process::Flags flags, RefPtr<Process>& outProcess);
-
-    static SlangResult getStdStream(Process::StreamType type, RefPtr<Stream>& out);
-
-        /// Sleep for time specified in milliseconds. 0 indicates to OS ok to yield this thread.
-    static void sleep(Index timeInMs);
-
-    static uint64_t getClockFrequency();
-
-    static uint64_t getClockTick();
 };
-
-// -----------------------------------------------------------------------
-SLANG_INLINE void CommandLine::addPrefixPathArg(const char* prefix, const String& path, const char* pathPostfix)
-{
-    StringBuilder builder;
-    builder << prefix << path;
-    if (pathPostfix)
-    {
-        // Work out the path with the postfix
-        builder << pathPostfix;  
-    }
-    addArg(builder.ProduceString());
-}
 
 }
